Added ability to insert multiple aliases:   - users/AddsAliases now accepts array of aliases;   - actions/CreateAliasDocument can overwrite previous aliases by same owner.

+++ src/actions/CreateAliasDocument.js
@@ -2,7 +2,6 @@
 
 const BaseAction = require('./BaseAction');
 const {AliasAlreadyExistsError} = require('../errors');
-const Db = require('../db/db');
 
 class CreateAliasDocument extends BaseAction {
   constructor (db, userId, alias) {
@@ -10,45 +9,53 @@ class CreateAliasDocument extends BaseAction {
     this.db = db;
     this.userId = userId;
     this.alias = alias;
+    this.docId = `alias:${this.alias.type}:${this.alias.value}`;
+    this.checkResult = null; // null if we don't have alias `docId`, doc itself if we do.
     this.result = null; // Couch reply to document craetion (ok, id, rev).
   }
 
   check (callback) {
-    const docId = `alias:${this.alias.type}:${this.alias.value}`;
-
-    this.db.exists(docId, (err, exists) => {
+    this.db.nullableGet(this.docId, (err, doc) => {
       if (err)
         return callback(err);
 
-      if (exists)
+      // existing alias, but owner is not userId
+      if (doc && (doc.id !== this.userId))
         return callback(new AliasAlreadyExistsError(this.alias.type, this.alias.value));
 
+      this.checkResult = doc;
       callback(null);
     });
   }
 
   execute (callback) {
-    const docId = `alias:${this.alias.type}:${this.alias.value}`;
-    const docBody = {
+    const docBody = Object.assign({}, this.checkResult || {}, {
       id: this.userId,
       date: this.alias.date,
-      public: this.alias.public
-    };
+      public: this.alias.public === true
+    });
 
-    this.db.save(docId, docBody, (err, response) => {
-      if (err instanceof Db.RevisionMismatchError)
-        return callback(new AliasAlreadyExistsError(this.alias.type, this.alias.value));
-
+    const cb = (err, response) => {
       if (err)
         return callback(err);
 
       this.result = response;
       callback(null);
-    });
+    };
+
+    return this.checkResult
+      ? this.db.replace(docBody, cb)
+      : this.db.save(this.docId, docBody, cb);
   }
 
   rollback (callback) {
-    this.db.delete(this.result.id, this.result.rev, callback);
+    // no previous alias
+    if (!this.checkResult)
+      return this.db.delete(this.result.id, this.result.rev, callback);
+
+    // restore original
+    const original = Object.assign({}, this.checkResult, {_rev: this.result.rev});
+    this.db.replace(original, callback);
   }
 }
